Default automatic biokit analysis

This wrapper function performs the "default" analyses of the package.

autoBiokitAnalysis(
  mat,
  sampInfo,
  groupCol,
  funCatList,
  statMode,
  annotationMode,
  filterByExpr,
  splitUpDown = TRUE,
  splitCol = "logFc",
  splitMidPoint = 0,
  sigCutoff = 0.05,
  metricCutoff = 1,
  rankCol = "logFc",
  topN = 50,
  decreasing = TRUE,
  absolute = TRUE,
  minFunCatSize = 1,
  maxFunCatSize = Inf,
  gseaRankCol = "logFc",
  gseaNPerm = 10000
)

Arguments

mat
The input matrix containing normalized data or counts

sampInfo
The sample information data frame.

groupCol
The column containing the grouping variable

funCatList
A list of character vectors with the functional categories to analyze.

statMode
Type of statistical analysis to perform the row-wise comparison. Allowed values are: "edgeR", "limma" "tTest", which will apply autoEdgerComparison, autoLimmaComparison and autoPairwiseMatrixTest, respectively.

annotationMode
Approach to annotate differential features. Allowed values are: "byCutoff or "byRank", which will apply addStatusByCutoff or addStatusByRank.

filterByExpr
If "edgeR" is selected, use filterByExpr?

splitUpDown
Split annotation into up and down features? If TRUE, splitCol and splitMidPoint should be provided.

splitCol
If splitUpDown is set to TRUE, the column used to split features into up and down.

splitMidPoint
If splitUpDown is set to TRUE, the splitCol value considered as midpoint.

sigCutoff
If annotation mode is "byCutoff", the adjusted p value cutoff to use.

metricCutoff
If annotation mode is "byCutoff", the absolute log FC cutoff to use.

rankCol
If annotation mode is "byRank", the column used to rank results.

topN
If annotation mode is "byRank", the top N features to annotate.

decreasing
If annotation mode is "byRank", order rankCol decreasing?

absolute
If annotation mode is "byRank", use absolute rankCol values?

minFunCatSize
Minimal size of a functional category to test. All functional categories below the threshold are excluded.

maxFunCatSize
Minimal size of a functional category to test. All functional categories aboce the threshold are excluded.

gseaRankCol
Column used to create pre-ranked gsea list.

gseaNPerm
Number of permutations to perform in the pre-ranked gsea analysis.

Value

A list containing three data frames: Statistical results, over-representation analysis results and gene set enrichment analysis results.