Obtain limma results from data, design and contrasts matrix.

Creates a linear model with limma and fits it to the provided contrasts. Then uses eBayes to smooth standard errors and obtain the list of pairwise comparison results with topTable.

limmaDfFromContrasts(
  mat,
  desMat,
  conMat,
  trend = FALSE,
  compName = "comparison",
  featName = "feature",
  exprName = "AveExpr",
  fcName = "logFc",
  pName = "pValue",
  pAdjName = "pAdj"
)

Arguments

mat
Matrix with data to analyse.

desMat
Design matrix.

conMat
Contrasts matrix.

trend
logical, should an intensity-trend be allowed for the prior variance? Default is that the prior variance is constant.

compName
Name for the column that indicates the pairwise comparison in the tidy data frame.

featName
Name for the column that indicates the analyzed row feature in the tidy data frame.

exprName
Name for the resulting average expresion column.

fcName
Name for the resulting fold change column.

pName
Name for the resulting p value column.

pAdjName
Name for the resulting adjusted p value column.

Value

A data frame with the results in a tidy format.