Generic single matrix analysis

Evaluates matrix data row-wise. This function is useful when the matrix consists on relative measures (i.e a matrix of log2 fold changes) or to analyze other features such as normal distribution.

singleMatrixTest(
  mat,
  metricFun = function(x) mean(x, na.rm = TRUE),
  statFun = nsTestT,
  sigFun = nsTestPValue,
  allowNa = TRUE,
  pAdjustMethod = "BH",
  featName = "feature",
  metricName = "logFc",
  statName = "t",
  pValName = "pValue",
  pAdjName = "pAdj"
)

Arguments

mat
Matrix containing data to be analyzed.

metricFun
Function to calculate metric of change. Defaults to log2(ratio of means).

statFun
Function to calculate statistic of change. Defaults to a non-sensitive T-Test T value.

sigFun
Function to calculate significance of change. Defaults to a non-sensitive T-Test P value.

allowNa
Allow NAs on input matrix?

pAdjustMethod
Method used to adjust P values.

featName
Name of the column containing the compared features (rows).

metricName
Name of the column containing the metric.

statName
Name of the column containing the statistic.

pValName
Name of the column containing the p values.

pAdjName
Name of the column containing the adjusted p values.

Value

A data frame with the row-wise statistical results.